feat: 파일을 제거하는 기능 추가.

+++ src/main/java/org/mayone/flow/model/FileDTO.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 @Data
 @Builder
 public class FileDTO {
-    private Long idx;
+    private int idx;
     private String originalName;
     private String saveName;
     private Long size;
